#1172, #1152, 1173: Improved API and DatasetMediaManager code
as per review, removed redundancies, shared code in common file.
