+++ signbank/api_interface.py
@@ -147,50 +147,6 @@ def get_gloss_data_json(request, datasetid, glossid):
     return JsonResponse(gloss_data)
 
 
-def get_gloss_filepath(video_file_path, gloss):
-
-    filename = os.path.basename(video_file_path)
-    filename_without_extension, _ = os.path.splitext(filename)
-    filepath, extension = os.path.splitext(video_file_path)
-    file_folder_path = os.path.dirname(video_file_path)
-    path_units = file_folder_path.split('/')
-    language_code_3char = path_units[-1]
-    dataset_acronym = path_units[-2]
-
-    # get language of import_videos path
-    language = Language.objects.filter(language_code_3char=language_code_3char).first()
-    if not language:
-        return "", "", ""
-    # get the annotation text of the gloss
-    annotationidglosstranslation = gloss.annotationidglosstranslation_set.all().filter(language=language)
-    if not annotationidglosstranslation:
-        # the gloss has no annotations for the language
-        return "", "", ""
-    annotation_text = annotationidglosstranslation.first().text
-    if annotation_text != filename_without_extension:
-        # gloss annotation does not match zip file name
-        return "", "", ""
-    two_letter_dir = get_two_letter_dir(gloss.idgloss)
-    destination_folder = os.path.join(
-        settings.WRITABLE_FOLDER,
-        settings.GLOSS_VIDEO_DIRECTORY,
-        dataset_acronym,
-        two_letter_dir
-    )
-
-    if not os.path.isdir(destination_folder):
-        os.mkdir(destination_folder)
-
-    glossid = str(gloss.id)
-
-    video_file_name = annotation_text + '-' + glossid + extension
-    goal = os.path.join(destination_folder, video_file_name)
-    video_path = os.path.join(settings.GLOSS_VIDEO_DIRECTORY,
-                              dataset_acronym,
-                              two_letter_dir)
-    return goal, video_file_name, video_path
-
-
 def check_gloss_existence_for_uploaded_video(dataset):
     dataset_acronym = str(dataset.acronym)
     goal_directory = os.path.join(VIDEOS_TO_IMPORT_FOLDER, dataset_acronym)
@@ -216,26 +172,6 @@ def check_gloss_existence_for_uploaded_video(dataset):
                         list_of_video_gloss_status[language3char].append((file, False, gloss))
 
     return list_of_video_gloss_status
-
-
-def uploaded_zip_archives(dataset):
-    # find the uploaded zip archives in TEMP that include the dataset
-    zip_archive = dict()
-    zipped_archive_directory = os.path.join(VIDEOS_TO_IMPORT_FOLDER, 'TEMP')
-    if os.path.isdir(zipped_archive_directory):
-        for file_or_folder in os.listdir(zipped_archive_directory):
-            file_or_folder_path = os.path.join(zipped_archive_directory, file_or_folder)
-            if os.path.isdir(file_or_folder_path):
-                continue
-            if not zipfile.is_zipfile(file_or_folder_path):
-                continue
-            archive_contents = get_filenames(file_or_folder_path)
-            common_prefix = os.path.commonprefix(archive_contents)
-            if dataset.acronym not in common_prefix:
-                continue
-            zipfilename = os.path.basename(file_or_folder_path)
-            zip_archive[zipfilename] = get_filenames_with_filetype(file_or_folder_path)
-    return zip_archive
 
 
 def uploaded_video_files(dataset):
@@ -410,80 +346,6 @@ def upload_zipped_videos_folder_json(request, datasetid):
     videos_data['unzippedvideos'] = unzipped_files
 
     return JsonResponse(videos_data)
-
-
-def remove_video_file_from_import_videos(video_file_path):
-    errors = ""
-    if not os.path.exists(video_file_path):
-        return errors
-    try:
-        os.remove(video_file_path)
-        errors = ""
-    except OSError:
-        errors = "Cannot delete the source video: " + video_file_path
-    return errors
-
-
-def get_two_letter_dir(idgloss):
-    foldername = idgloss[:2]
-
-    if len(foldername) == 1:
-        foldername += '-'
-
-    return foldername
-
-
-def save_video(video_file_path, goal):
-    # this is called inside an atomic block
-
-    try:
-        shutil.copyfile(video_file_path, goal)
-        return True
-    except IOError:
-        return False
-
-
-def import_video_file(request, gloss, video_file_path):
-    # request is included as a parameter to add to the GlossVideoHistory
-    with atomic():
-        goal_gloss_file_path, video_file_name, video_path = get_gloss_filepath(video_file_path, gloss)
-        if not goal_gloss_file_path:
-            return "Failed", "Incorrect gloss path for import"
-        existing_videos = GlossVideo.objects.filter(gloss=gloss)
-        if existing_videos.count():
-            for video_object in existing_videos:
-                video_object.reversion(revert=False)
-
-        # overwritten should not happen because we already backed up the original videos
-        success = save_video(video_file_path, goal_gloss_file_path)
-
-        if success:
-            # make new GlossVideo object for new video
-            video = GlossVideo(gloss=gloss,
-                               version=0)
-            with open(goal_gloss_file_path, 'rb') as f:
-                f.seek(0)
-                video.videofile.save(os.path.basename(goal_gloss_file_path), File(f), save=False)
-            video.save()
-            video.make_poster_image()
-            glossvideohistory = GlossVideoHistory(action="import",
-                                                  gloss=gloss,
-                                                  actor=request.user,
-                                                  uploadfile=video_file_path,
-                                                  goal_location=goal_gloss_file_path)
-            glossvideohistory.save()
-            status = 'Success'
-        else:
-            # import failed to copy new video, put originals back
-            if existing_videos.count():
-                for video_object in existing_videos:
-                    video_object.reversion(revert=True)
-            status = 'Failed'
-
-        # errors are if the import_videos video can not be removed
-        # errors = remove_video_file_from_import_videos(video_file_path)
-        errors = ""
-        return status, errors
 
 
 def import_video_to_gloss(request, video_file_path):
